requirements.txt
boolean.py
httpx
llama-parse
pdf2image
pillow
pymongo
pymupdf
python-dotenv
pytesseract
requests
setuptools
streamlit
streamlit-aggrid
opencv-python-headless
wheel
streamlit
openai
numpy
chromadb
python-dotenv
azure-identity
azure-core
weasyprint
python-docx
docxtpl
